+++ jspp/js/modules/widgets/Radio.js
@@ -22,7 +22,7 @@ class Radio extends lx.Checkbox {
 	}
 
 	static initCss(css) {
-		css.useHolder(lx.BasicCssContext);
+		css.useExtender(lx.BasicCssContext);
 		css.addClass('lx-Radio-0', {
 			border: 'solid #61615e 1px',
 			width: '16px',