+++ jspp/js/src/common/tools/makeup/PresetFieldtHolder.js
@@ -1,39 +1,4 @@
 // @lx:namespace lx;
-class CssContextHolder {
-    static context;
-
-    /**
-     * @abstract 
-     * @param {lx.CssContext} css
-     */
-    static init(css) {
-        // pass
-    }
-
-    /**
-     * @returns lx.CssContext
-     */
-    static getContext() {
-        if (!this.context) {
-            const css = new lx.CssContext;
-            this.init(css);
-            this.context = css;
-        }
-        return this.context;
-    }
-
-    /**
-     * @param {string|Array<string>} valName
-     * @param {any|Array<any>} defaultVal
-     * @param {Function} [modifier]
-     * @returns {lx.PresetFieldtHolder}
-     */
-    static presetValue(valName, defaultVal, modifier = null) {
-        return new lx.PresetFieldtHolder(valName, defaultVal, modifier);
-    }
-}
-
-// @lx:namespace lx;
 class PresetFieldtHolder {
     /**
      * @param {string|Array<string>} valName